requirements.txt
cores==0.4.1
dataclasses_json==0.6.2
discord.py==2.3.2
python-dotenv==1.0.0
PyYAML==6.0.1
supabase==2.0.3
